\name{mcif-package}
\alias{mcif-package}
\alias{mcif}
\docType{package}
\title{
\packageTitle{mcif}
}
\description{
\packageDescription{mcif}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mcif}
\packageIndices{mcif}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mcif}

Maintainer: \packageMaintainer{mcif}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
